Create PowerShell script that detects privileged accounts in Active Directory
This script uses the Get-ADUser cmdlet to search for privileged accounts in the specified domain or organizational unit. The search filter looks for users who are members of either the Domain Admins or Enterprise Admins groups. The Select-Object cmdlet is used to retrieve the user's name, SamAccountName (i.e. the user's login name), and the groups that the user is a member of. Finally, the results are sorted by name and formatted as a table.

Note that you'll need to run this script with appropriate permissions to read Active Directory information.