Create PowerShell script that detects local accounts in a device
This script uses the Get-WmiObject cmdlet to search for local accounts on the specified computer. The Where-Object cmdlet is used to filter the results to only include local accounts (i.e. accounts with the LocalAccount property set to $true). The Select-Object cmdlet is used to retrieve the account's name, caption (i.e. the user's full name), and account type (e.g. user or administrator). Finally, the results are sorted by name and formatted as a table.

Note that you'll need to run this script with appropriate permissions to access the WMI information. Also, some newer versions of PowerShell recommend using the Get-CimInstance cmdlet instead of Get-WmiObject. So, you may consider modifying the script based on your PowerShell version.