Create PowerShell script that detects local accounts on all devices in a given Active Directory domain
This script uses the Get-ADComputer cmdlet to retrieve all domain computers in the specified domain. It then loops through each computer and uses the Get-WmiObject cmdlet to search for local accounts on that computer. The Where-Object cmdlet is used to filter the results to only include local accounts (i.e. accounts with the LocalAccount property set to $true). The Select-Object cmdlet is used to retrieve the account's name, caption (i.e. the user's full name), account type (e.g. user or administrator), and the name of the computer that the account was found on. Finally, the results are sorted by computer name and account name, and formatted as a table.

Note that you'll need to run this script with appropriate permissions to access the WMI information on each computer. Additionally, this script assumes that all computers are in the same domain and that you have the appropriate permissions to access Active Directory information.